+++ changelog.md
@@ -7,6 +7,7 @@ See also the commit log at github: https://github.com/Arduino-IRremote/Arduino-I
 - Fixed error with DEBUG in TinyIRReceiver.hpp.
 - Support for ATmega88 see issue #923. Thanks to Dolmant.
 - NO_LED_FEEDBACK_CODE replaces and extends DISABLE_LED_FEEDBACK_FOR_RECEIVE.
+- Removed NO_LEGACY_COMPATIBILITY macro, it was useless now.
 
 ## 3.5.0
 - Improved ir_DistanceProtocol.